chain: use reorganize() instead of _replay() for _invalidate()
The benefit of using reorganize() is that it disconnects blocks
individually from the tip down to (and including) the invalid block.
This will cause chain to emit a 'disconnect' event for each removed
block, and finish with a 'reorganize' event. The 'reorganize' event
is important because it signals to the mempool to check for and
evict transactions which are no longer valid due to CSV locks or
coinbase maturity.